Additions to Events.cpp and Events.h
Finished method for addTimeSlots in Events.cpp and added a destructor that deletes LinkedList of time slots when events class is deleted. Also wrote operator overloads for >, <, == to compare for year, month and day, though they still need to comapre time slots. So the operator overloads are still unfinished. Also corrected a sspelling error in TimeSlots.cpp.