Finished Events >,<, == operator overloads... sort of
The >,<, == operator overloads compare correctly for year, month and day. When comparing time slots, they only compare the starting times for events.  I'm not sure if that is what's wanted. Need to discuss this further, probably will need to alter that somewhat. Besides that, also got rid of s_time and e_time variables.