Comments
Added comments to all header files and started to add comments to the .cpp files as well. Executive is only half commented but the rest of the .cpps are finished.
Modified some of the code to make it simpler but there is much more work to be done with rewriting code.