Templated LinkedList
LinkedList is now free from any type specific methods. It is now double templeted with a keyType (which cooresponds to the type held in a node) and a searchType (which allows you to search the list with a different type).
I also hopefully integrated it with the rest of the program so it was as if nothing changed. It did however lead to the TimeSlot list in Event to have keyType == searchType so we may want to revisit that later.